LayoutTests/imported/w3c/web-platform-tests/cors/redirect-userinfo-expected.txt

Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?&location=http%3A%2F%2Ftest%3Atest%40www1.localhost%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3F%26get_value%3Dlast%260_1&0_0
CONSOLE MESSAGE: line 57: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?&location=http%3A%2F%2Ftest%3Atest%40www1.localhost%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3F%26get_value%3Dlast%260_1&0_0 due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?&location=http%3A%2F%2Fuser%3A%40www1.localhost%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3F%26get_value%3Dlast%261_1&1_0
CONSOLE MESSAGE: line 57: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?&location=http%3A%2F%2Fuser%3A%40www1.localhost%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3F%26get_value%3Dlast%261_1&1_0 due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?&location=http%3A%2F%2Fuser%3A%40www1.localhost%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3F%26get_value%3Dlast%262_1&2_0
CONSOLE MESSAGE: line 57: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?&location=http%3A%2F%2Fuser%3A%40www1.localhost%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3F%26get_value%3Dlast%262_1&2_0 due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?&location=http%3A%2F%2F%3A%40www1.localhost%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3F%26get_value%3Dlast%263_1&3_0
CONSOLE MESSAGE: line 77: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?&location=http%3A%2F%2F%3A%40www1.localhost%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3F%26get_value%3Dlast%263_1&3_0 due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?&location=http%3A%2F%2F%3Apass%40www1.localhost%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3F%26get_value%3Dlast%264_1&4_0
CONSOLE MESSAGE: line 57: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?&location=http%3A%2F%2F%3Apass%40www1.localhost%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3F%26get_value%3Dlast%264_1&4_0 due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?&location=http%3A%2F%2F%40www1.localhost%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3F%26get_value%3Dlast%265_1&5_0
CONSOLE MESSAGE: line 77: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?&location=http%3A%2F%2F%40www1.localhost%3A8800%2Fcors%2Fresources%2Fcors-makeheader.py%3F%26get_value%3Dlast%265_1&5_0 due to access control checks.
CORS userinfo redirect handling


PASS Disallow redirect with userinfo (user:pass@) 
PASS Disallow redirect with userinfo (user:@) 
PASS Disallow redirect with userinfo (user@) 
FAIL Allow redirect without userinfo (:@ is trimmed during URL parsing) assert_unreached: Reached unreachable code
PASS Disallow redirect with userinfo (:pass@) 
FAIL Allow redirect without userinfo (@ is trimmed during URL parsing) assert_unreached: Reached unreachable code
